+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog Natrium
 
+## v1.3 (15-06-2016)
+- Added `appicon` directive
+
+## v1.2 (09-06-2016)
+- Added `files` directive
+
 ## v1.1 (12-05-2016)
 - Easier setup with CocoaPods
 - Multiple platforms (iOS / OSX)